Fix little english mistakes + Moved cache decorator to the root folder instead of 'routes' and edited connection message.
