+++ db/decorator.go
@@ -26,7 +26,7 @@ import (
 )
 
 // Transaction is a decorator which manages a transaction for an HTTP request.
-// It will Commit the transaction iff the handler returns something other than
+// It will Commit the transaction if the handler returns something other than
 // an error and it did not panic; it Rollbacks otherwise.
 type RequestTransaction struct {
 	Db *sql.DB
