+++ usageReports/es/es_route.go
@@ -16,7 +16,7 @@ package es
 
 import (
 	"database/sql"
-	"github.com/trackit/trackit-server/routes/cache"
+	"github.com/trackit/trackit-server/cache"
 	"net/http"
 	"time"
